smbios: nuke the XLEN fields
SMBIOS is used to describe physical assets of a system. For a
processor this may include uArch/SoC/vendor versioning, but the
XLEN fields themselves aren't useful - and if they are, they belong
in other specs that are critical to OS startup.

As mentioned previously, SMBIOS struct len is a bit of a premium
so there needs to be a great reason to include particular info.

Signed-off-by: Andrei Warkentin <andrei.warkentin@intel.com>