Make PCI initialization optional for BRS-B
End-users have had trouble over the years with requiring BIOS to
configure PCIe. On the System 76 laptop with 4 thunderbolt
ports, it was an issue; and, on hyperscalar server platforms,
there have been cases where it is best to let linux configure
pcie, as firmware does not optimize to suit the case. The
system-76 case was a case also seen years ago on HPC nodes: BIOS
was required to configure the system for maximum generality,
making certain hardware impossible to use (4th thunderbolt port
on system 76; Quadrics NICs on HPC); in each case, the fix was
to get the BIOS to do minimal PCI configuration. Coreboot has a
build time option to enable this minimal PCI configuration,
deveoped jointly by Google and Intel.

Signed-off-by: Ronald G Minnich <rminnich@gmail.com>