smbios: update processor ID for RISCV to reflect the SoC ID
Implements https://github.com/riscv-non-isa/riscv-brs/issues/147

The way this is done, it's forwards compatible with the DMTF
spec text, which should help with upstreaming. Compatible
in the sense that the "old" way just provided the JEP-106
value in the first DWORD (and had 0 for the second DWORD).

Note: unlike Arm, don't bother with a processor characteristic,
as there is no analogue to the SMCCC SoC ID call.

Signed-off-by: Andrei Warkentin <andrei.warkentin@intel.com>